feat : friendship controllers/services

+++ src/friendships/friendships.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Get, Request, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import { Controller, Get, Request, Post, Body, Patch, Param, Delete, HttpException, HttpStatus } from '@nestjs/common';
 import { FriendshipsService } from './friendships.service';
 import { CreateFriendshipDto } from './dto/create-friendship.dto';
 import { UpdateFriendshipDto } from './dto/update-friendship.dto';
@@ -6,6 +6,7 @@ import { ApiProperty, ApiTags } from "@nestjs/swagger";
 import { UsersService } from 'src/users/users.service';
 import { UseGuards } from '@nestjs/common';
 import { JwtAuthGuard } from 'src/auth/jwt-auth.guard';
+import { ParseIntPipe } from '@nestjs/common/pipes';
 
 
 @ApiTags('friendships')
@@ -20,27 +21,50 @@ export class FriendshipsController {
   async createFriendship(@Body() body: CreateFriendshipDto, @Request() req) {
     const user = await this.usersService.findUserById(req.user.userId);
     const friend = await this.usersService.findUserByPseudo(body.pseudo);
-
     return await this.friendshipsService.createFriendship(user, friend);
   }
 
-  @Get()
-  findAll() {
-    return this.friendshipsService.findAll();
-  }
-
+  @UseGuards(JwtAuthGuard)
   @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.friendshipsService.findOne(+id);
+  async findById(@Param('id', ParseIntPipe) id: number) {
+    const friendship = await this.friendshipsService.findOne(id);
+    if (!friendship) {
+      throw new HttpException("Relation introuvable", HttpStatus.NOT_FOUND);
+    }
+    return friendship;
   }
 
+  @UseGuards(JwtAuthGuard)//accepte l'invitation et sert à créer la relation inverse de friendId vers UserId
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateFriendshipDto: UpdateFriendshipDto) {
-    return this.friendshipsService.update(+id, updateFriendshipDto);
+  async update(@Param('id', ParseIntPipe) id: number, @Request() req) {
+    const friendship = await this.friendshipsService.findOne(id);
+    if (!friendship) {
+      throw new HttpException(" realtion introuvable", HttpStatus.NOT_FOUND);
+    }
+    const user = await this.usersService.findUserById(req.user.userId);
+    
+    if (friendship.friend.id !== user.id) {
+      throw new HttpException(" Non autorisé.", HttpStatus.FORBIDDEN);
+    }
+    if (friendship.accepted) {
+      throw new HttpException("Already accepted .", HttpStatus.BAD_REQUEST);
+    }
+    await this.friendshipsService.update(id);
+    const inverseRelation = await this.friendshipsService.createFriendship(friendship.friend, friendship.user);
+    return await this.friendshipsService.update(inverseRelation.id);
+    
   }
 
+  @UseGuards(JwtAuthGuard)
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  async remove(@Param('id', ParseIntPipe) id: number, @Request() req) {
+    const friendship = await this.friendshipsService.findOne(id);
+    if (!friendship) {
+      throw new HttpException(" relation introuvable", HttpStatus.NOT_FOUND);
+    }
+    if (friendship.friend.id !== user.id) {
+      throw new HttpException(" Non autorisé.", HttpStatus.FORBIDDEN);
+    }
     return this.friendshipsService.remove(+id);
   }
 }
